+++ test/generate_rda.R
@@ -55,3 +55,29 @@ save(test.cmpfun0, test.cmpfun1, test.cmpfun2, file = "data/cmpfun.rda")
 x <- factor(c("a", "b", "c"))
 y <- ordered(x, levels=c("b", "a", "c"))
 save(x, y, file="data/ord.rda")
+
+dates = as.Date("2017-01-01") + 1:4
+datetimes = as.POSIXct("2017-01-01 13:23", tz="UTC") + 1:4
+dateNAs = list(c(dates, NA), c(datetimes, NA))
+saveRDS(dateNAs, file="data/datesNA.rds")
+datelst = list(dates, dates[1])
+names(dates) = LETTERS[1:length(dates)]
+datelst = c(datelst, list(dates), list(dates[1]))
+saveRDS(datelst, file="data/dates.rds")
+dtlst = list(datetimes, datetimes[1])
+names(datetimes) = LETTERS[1:length(datetimes)]
+dtlst = c(dtlst, list(datetimes), list(datetimes[1]))
+saveRDS(dtlst, file="data/datetimes.rds")
+datedfs = list(data.frame(date=dates[1], datetime=datetimes[1]),
+               data.frame(date=dates, datetime=datetimes))
+saveRDS(datedfs, file="data/datedfs.rds")
+
+# the first element here is assumed to be in the local timezone but is saved in
+# UTC time, without any timezone attribute. When R reads it, it assumes local time.
+# So the test associated with this first datapoint is going to assume which timezone
+# the data is generated in! (PST/-8)
+saveRDS(list(as.POSIXct("2017-01-01 13:23"),
+             as.POSIXct("2017-01-01 13:23", tz="CST"),
+             as.POSIXct("2017-01-01 13:23", tz="America/Chicago")),
+        file="data/datetimes_tz.rds")
+